Extend and rename RUF004 to PLR1722
