+++ resources/test/fixtures/pylint/consider_using_sys_exit_1.py
@@ -1,10 +1,12 @@
 import sys
 
 exit(0)
+quit(0)
 
 
 def main():
     exit(1)
+    quit(1)
 
 
 sys.exit(2)
