+++ resources/test/fixtures/pylint/consider_using_sys_exit_4.py
@@ -1,7 +1,9 @@
 from sys import exit as exit2
 
 exit(0)
+quit(0)
 
 
 def main():
     exit(1)
+    quit(1)
