+++ resources/test/fixtures/pylint/consider_using_sys_exit_6.py
@@ -1,0 +1,19 @@
+exit(0)
+quit(0)
+
+
+def exit(e):
+    pass
+
+
+def quit(e):
+    pass
+
+
+exit(1)
+quit(1)
+
+
+def main():
+    exit(2)
+    quit(2)
