--- resources/test/fixtures/ruff/RUF004_0.py
@@ -1,5 +1,0 @@
-exit(0)
-
-
-def main():
-    exit(2)
