+++ src/checks_gen.rs
@@ -309,6 +309,8 @@ pub enum CheckCodePrefix {
     PLR17,
     PLR170,
     PLR1701,
+    PLR172,
+    PLR1722,
     PLW,
     PLW0,
     PLW01,
@@ -338,7 +340,6 @@ pub enum CheckCodePrefix {
     RUF001,
     RUF002,
     RUF003,
-    RUF004,
     RUF1,
     RUF10,
     RUF100,
@@ -1229,9 +1230,12 @@ impl CheckCodePrefix {
             CheckCodePrefix::PLE11 => vec![CheckCode::PLE1142],
             CheckCodePrefix::PLE114 => vec![CheckCode::PLE1142],
             CheckCodePrefix::PLE1142 => vec![CheckCode::PLE1142],
-            CheckCodePrefix::PLR => {
-                vec![CheckCode::PLR0206, CheckCode::PLR0402, CheckCode::PLR1701]
-            }
+            CheckCodePrefix::PLR => vec![
+                CheckCode::PLR0206,
+                CheckCode::PLR0402,
+                CheckCode::PLR1701,
+                CheckCode::PLR1722,
+            ],
             CheckCodePrefix::PLR0 => vec![CheckCode::PLR0206, CheckCode::PLR0402],
             CheckCodePrefix::PLR02 => vec![CheckCode::PLR0206],
             CheckCodePrefix::PLR020 => vec![CheckCode::PLR0206],
@@ -1239,10 +1243,12 @@ impl CheckCodePrefix {
             CheckCodePrefix::PLR04 => vec![CheckCode::PLR0402],
             CheckCodePrefix::PLR040 => vec![CheckCode::PLR0402],
             CheckCodePrefix::PLR0402 => vec![CheckCode::PLR0402],
-            CheckCodePrefix::PLR1 => vec![CheckCode::PLR1701],
-            CheckCodePrefix::PLR17 => vec![CheckCode::PLR1701],
+            CheckCodePrefix::PLR1 => vec![CheckCode::PLR1701, CheckCode::PLR1722],
+            CheckCodePrefix::PLR17 => vec![CheckCode::PLR1701, CheckCode::PLR1722],
             CheckCodePrefix::PLR170 => vec![CheckCode::PLR1701],
             CheckCodePrefix::PLR1701 => vec![CheckCode::PLR1701],
+            CheckCodePrefix::PLR172 => vec![CheckCode::PLR1722],
+            CheckCodePrefix::PLR1722 => vec![CheckCode::PLR1722],
             CheckCodePrefix::PLW => vec![CheckCode::PLW0120],
             CheckCodePrefix::PLW0 => vec![CheckCode::PLW0120],
             CheckCodePrefix::PLW01 => vec![CheckCode::PLW0120],
@@ -1312,25 +1318,13 @@ impl CheckCodePrefix {
                 CheckCode::RUF001,
                 CheckCode::RUF002,
                 CheckCode::RUF003,
-                CheckCode::RUF004,
                 CheckCode::RUF100,
             ],
-            CheckCodePrefix::RUF0 => vec![
-                CheckCode::RUF001,
-                CheckCode::RUF002,
-                CheckCode::RUF003,
-                CheckCode::RUF004,
-            ],
-            CheckCodePrefix::RUF00 => vec![
-                CheckCode::RUF001,
-                CheckCode::RUF002,
-                CheckCode::RUF003,
-                CheckCode::RUF004,
-            ],
+            CheckCodePrefix::RUF0 => vec![CheckCode::RUF001, CheckCode::RUF002, CheckCode::RUF003],
+            CheckCodePrefix::RUF00 => vec![CheckCode::RUF001, CheckCode::RUF002, CheckCode::RUF003],
             CheckCodePrefix::RUF001 => vec![CheckCode::RUF001],
             CheckCodePrefix::RUF002 => vec![CheckCode::RUF002],
             CheckCodePrefix::RUF003 => vec![CheckCode::RUF003],
-            CheckCodePrefix::RUF004 => vec![CheckCode::RUF004],
             CheckCodePrefix::RUF1 => vec![CheckCode::RUF100],
             CheckCodePrefix::RUF10 => vec![CheckCode::RUF100],
             CheckCodePrefix::RUF100 => vec![CheckCode::RUF100],
@@ -1966,6 +1960,8 @@ impl CheckCodePrefix {
             CheckCodePrefix::PLR17 => SuffixLength::Two,
             CheckCodePrefix::PLR170 => SuffixLength::Three,
             CheckCodePrefix::PLR1701 => SuffixLength::Four,
+            CheckCodePrefix::PLR172 => SuffixLength::Three,
+            CheckCodePrefix::PLR1722 => SuffixLength::Four,
             CheckCodePrefix::PLW => SuffixLength::Zero,
             CheckCodePrefix::PLW0 => SuffixLength::One,
             CheckCodePrefix::PLW01 => SuffixLength::Two,
@@ -1995,7 +1991,6 @@ impl CheckCodePrefix {
             CheckCodePrefix::RUF001 => SuffixLength::Three,
             CheckCodePrefix::RUF002 => SuffixLength::Three,
             CheckCodePrefix::RUF003 => SuffixLength::Three,
-            CheckCodePrefix::RUF004 => SuffixLength::Three,
             CheckCodePrefix::RUF1 => SuffixLength::One,
             CheckCodePrefix::RUF10 => SuffixLength::Two,
             CheckCodePrefix::RUF100 => SuffixLength::Three,
