--- src/rules/plugins/mod.rs
@@ -1,3 +1,0 @@
-pub use convert_exit_to_sys_exit::convert_exit_to_sys_exit;
-
-mod convert_exit_to_sys_exit;
